% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific_validation.R
\name{bhf_collect_validation_targets}
\alias{bhf_collect_validation_targets}
\title{Collect Scientific Target Draws}
\usage{
bhf_collect_validation_targets(fit)
}
\description{
Collect Scientific Target Draws
}
\keyword{internal}
