% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g4r_validation.R
\name{bhf_g4r_candidate_seeds}
\alias{bhf_g4r_candidate_seeds}
\title{Map a Prospective Candidate Index to Frozen Seeds}
\usage{
bhf_g4r_candidate_seeds(profile, candidate_index)
}
\description{
Map a Prospective Candidate Index to Frozen Seeds
}
\keyword{internal}
