% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g4r_validation.R
\name{bhf_g4r_gap_metrics}
\alias{bhf_g4r_gap_metrics}
\title{Calculate the Frozen G4-R Gap Statistic}
\usage{
bhf_g4r_gap_metrics(estimands, protocol = bhf_g4r_protocol())
}
\description{
Calculate the Frozen G4-R Gap Statistic
}
\keyword{internal}
