% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g4r_validation.R
\name{bhf_g4r_weight_arm}
\alias{bhf_g4r_weight_arm}
\title{Materialize the Frozen Nondegenerate G4-R Weight Arm}
\usage{
bhf_g4r_weight_arm(truth_only, protocol = bhf_g4r_protocol())
}
\description{
Materialize the Frozen Nondegenerate G4-R Weight Arm
}
\keyword{internal}
