% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{bhf_plot_astar_sensitivity}
\alias{bhf_plot_astar_sensitivity}
\title{Plot the A* Sampling-Variance Sensitivity}
\usage{
bhf_plot_astar_sensitivity(x, what = c("proportion", "between"))
}
\arguments{
\item{x}{The result of \code{\link[=bhf_astar_sensitivity]{bhf_astar_sensitivity()}}.}

\item{what}{Which quantity to plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the A* Sampling-Variance Sensitivity
}
\examples{
\dontrun{
bhf_plot_astar_sensitivity(bhf_astar_sensitivity(fit))
}
}
\seealso{
\code{\link[=bhf_astar_sensitivity]{bhf_astar_sensitivity()}}
}
