% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific_validation.R
\name{bhf_scientific_sensitivity_grid}
\alias{bhf_scientific_sensitivity_grid}
\title{Scientific Sensitivity Grid}
\usage{
bhf_scientific_sensitivity_grid()
}
\description{
Scientific Sensitivity Grid
}
\keyword{internal}
