% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific_validation.R
\name{bhf_validation_artifact_status}
\alias{bhf_validation_artifact_status}
\title{Check Whether a Validation Artifact Can Be Resumed}
\usage{
bhf_validation_artifact_status(path, config_hash, expected = list())
}
\description{
Check Whether a Validation Artifact Can Be Resumed
}
\keyword{internal}
