% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{calc_eff_n}
\alias{calc_eff_n}
\title{Calculate Effective Sample Size}
\usage{
calc_eff_n(weights)
}
\arguments{
\item{weights}{Numeric vector of survey weights.}
}
\value{
Effective sample size (scalar).
}
\description{
Calculates the effective sample size given survey weights.
}
\examples{
w <- c(1, 2, 1.5, 3, 2)
calc_eff_n(w)  # Should be less than length(w)

}
