% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhf_fit.R
\name{compute_diagnostics}
\alias{compute_diagnostics}
\title{Compute MCMC Diagnostics}
\usage{
compute_diagnostics(stanfit)
}
\arguments{
\item{stanfit}{A stanfit object.}
}
\value{
List of diagnostic summaries.
}
\description{
Internal function to compute diagnostic summaries from a stanfit object.
}
\keyword{internal}
